Split SymbolicValue out to being defined at the SIL level, in prep for Dan's GraphOperationInst.  Also, make the SerializedSILLoader be lazily allocated on first use, since it is expensive to create.
